+++ src/QueryCat.Backend/Application.cs
@@ -1,3 +1,5 @@
+using Microsoft.Extensions.Logging;
+using Microsoft.Extensions.Logging.Abstractions;
 using QueryCat.Backend.Functions.StandardFunctions;
 
 namespace QueryCat.Backend;
@@ -5,8 +7,13 @@ namespace QueryCat.Backend;
 /// <summary>
 /// Application information.
 /// </summary>
-public static class QueryCatApplication
+public static class Application
 {
+    /// <summary>
+    /// Default application log factory.
+    /// </summary>
+    public static ILoggerFactory LoggerFactory { get; set; } = NullLoggerFactory.Instance;
+
     /// <summary>
     /// Product name.
     /// </summary>