+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Changed
 
 - Class renames: ClassEnumerableInput to FetchInput, ClassEnumerableInputFetch to Fetcher.
+- Change logger to Microsoft.Extensions.Logging.
 
 ## [0.3.5] - 2023-03-26